[droideglsink] Rework rendering
We now do it in a way similar to Android.

Application needs to call attach_to_display() passing an EGLDisplay object and call
detach_from_display() in response to frame-ready signal with -1 as the frame number

We will also destroy the previous fence when we try to acquire the new frame in order to give the GL
implementation enough time to finish rendering.

The buffer is held by us too until this is done via an extra ref